Autosuggest: Tweak the selector where the results are injected into.
Previously, if you register multiple "resultsbox" areas on the same
page, there will be multiple #mentions-autosuggest UL elements created.

When an AJAX request is made, the JS would only inject the results
into the first occurence of the #mentions-autosuggest UL.

This commit adds a new javascript "parent" variable to calculate the
parent of the resultsbox.  Then, we try to find the
right below the proper textarea.